Fixed coding standard violations. Fixed identation, placed Uut in a seperate class file
